Split invitation service into 3 services
